requirements.txt
pandas==2.0.3
python-dotenv==1.0.1
azure-identity==1.14.0
azure-storage-blob==12.18.3
pyarrow==14.0.2
